# Optimize Enumerable#join