This is an insignificant PR; I noticed this when working on a lexer/builder.

```cr
require "benchmark"

module Enumerable
  def new_join(separator, io)
    start = false
    each do |elem|
      start ? io << separator : (start = true)
      yield elem, io
    end
  end
end

SMALL = StaticArray(Int32, 4).new 123
BIG = StaticArray(Int32, 999).new 123

Benchmark.ips do |x|
  x.report("small old join") { SMALL.join("", STDOUT) {|e|} }
  x.report("small new join") { SMALL.new_join("", STDOUT) {|e|} }

  x.report("big old join") { BIG.join("", STDOUT) {|e|} }
  x.report("big new join") { BIG.new_join("", STDOUT) {|e|} }
end
```
Result:
```
small old join  73.28M ( 13.65ns) (± 0.81%)  0 B/op         fastest
small new join  72.57M ( 13.78ns) (± 2.34%)  0 B/op    1.01× slower
  big old join 226.14k (  4.42µs) (± 1.16%)  0 B/op  324.02× slower
  big new join 231.49k (  4.32µs) (± 1.36%)  0 B/op  316.54× slower
```
 
Edit My bad I mess up the condition, it wasn't doing anything. Fixed and result updated.
Using booleans are easier to understand.