@asterite 2019-04-25
Could you write it like this?

```crystal
  def join(separator, io : IO)
    first = true
    each_with_index do |elem, i|
      io << separator unless first
      yield elem, io
      first = false
    end
  end
```

Doing an assignment inside a ternary operator is very hard to read and understand.