@asterite 2019-04-25
There's no magic here: the existing code increments a counter on each loop. The new code just sets a variable to false. I guess that's slightly faster.