@j8r 2019-04-25
Ok, but the `each_wth_index` is no longer needed @asterite