@konovod 2019-04-26
So there is no performance bonus in benchmark with latest changes? What's a point of change then?