@j8r 2019-04-26
@konovod there is a slight difference when using bigger enumerable:
```
big old join 202.45k (  4.94µs) (± 3.37%)  0.0B/op   1.01× slower
big new join 204.20k (  4.90µs) (± 2.88%)  0.0B/op        fastest
```
I reproduced it several time to be sure.
I admit this is a bit useless, I was initially mainly wondering if there is an advantage of incrementing a counter compared to a simple boolean.