@straight-shoota 2019-04-26
Did you check if moving `first = false` into the `unless first` branch provides any notable benefit?

```cr
first = true
each do |elem|
  unless first
    io << separator
    first = false
  end
  yield elem, io
end
```