@j8r 2019-04-26
@straight-shoota the code inside the unless will never run.
I've checked
```cr
first = true
each do |elem|
  if first
    first = false
  else
    io << separator
  end
  yield elem, io
end
```
No difference compared to doing `first = false` on each iteration (@asterite code). Maybe it doesn't necessarily means an operation if `first` is already `false`, or it is as expensive as a `if/else` conditional.