@wontruefree 2019-05-01
@j8r I did a few benchmarks and found that code you mentioned had some benefits with big arrays (~ 3.5%).

But I found that checking the variable before writing back to it has about a ~10% benefit with big arrays (9999999 elements) and about a ~10% loss with small arrays (99 elements).

```crystal
  def new_join(separator, io)
    first = true
    each do |elem|
      io << separator unless first
      yield elem, io
      first = false if first
    end
  end
```

Don't know if it is worth it because I imagine most joins will be with small arrays.  Also, I am not entirely sure why this happens but I was thinking setting the variable would break the CPU cache and that would cause a performance hit.