@wontruefree 2019-05-01
@konovod I would agree with you which is why I am stating I don't know why this is happening.  From your benches, I am not sure which strategy is best.  It seems like they are all the fastest at some point.