@jhass 2019-05-23
The benchmark results are essentially random and within error margin towards each other. The optimizer is smart enough here and this is bikeshedding. I'm in favor of closing this as there's no clear evidence of improvement.